# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/mine": {
      "get": {
        "operationId": "EventsController_findMine",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "PENDING_MODERATION",
                "PUBLISHED",
                "REJECTED",
                "CANCELLED",
                "COMPLETED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}/publish": {
      "post": {
        "operationId": "EventsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/slug/{slug}": {
      "get": {
        "operationId": "EventsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kiro API",
    "description": "Backend API for the Kiro events platform",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "uk",
              "en"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "format": {
            "type": "string",
            "enum": [
              "OFFLINE",
              "ONLINE"
            ]
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "cityId": {
            "type": "string"
          },
          "districtId": {
            "type": "string"
          },
          "addressText": {
            "type": "string"
          },
          "addressDetails": {
            "type": "object",
            "description": "Structured address components from Google Places (§60)."
          },
          "googlePlaceId": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "onlineUrl": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "minParticipants": {
            "type": "number"
          },
          "registrationDeadline": {
            "type": "string"
          },
          "approvalMode": {
            "type": "string",
            "enum": [
              "AUTO",
              "ORGANIZER_APPROVAL"
            ]
          },
          "ageRestriction": {
            "type": "number"
          },
          "rules": {
            "type": "string"
          },
          "priceType": {
            "type": "string",
            "enum": [
              "FREE",
              "PAID"
            ]
          },
          "price": {
            "type": "number"
          },
          "paymentUrl": {
            "type": "string"
          },
          "notifyParticipants": {
            "type": "boolean",
            "description": "Confirms the organizer wants participants notified of a significant change (§78). Required to be true when changing startsAt/address/onlineUrl on an already-published event."
          }
        }
      },
...
    }
  }
}